pyproject.toml
[project]
name = "01-eprompt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "guardrails-ai==0.6.0",
    "guardrails-api-client==0.4.0a1",
    "pydantic>=2.10.2",
    "python-dotenv>=1.0.1",
]
